requirements.txt
python-dotenv==1.0.0
pymongo==4.5.0
openai==0.28.1
pydantic==2.4.2
fastapi==0.103.2
uvicorn==0.23.2